Function algorithms::merge_sort

pub fn merge_sort<A: Ord + Clone>(data: &[A]) -> Vec<A>

Efficient sorting against large sets. Requires O(n) aux. space.

This divide-and-conquer sorting algorithm, while inefficient with memory use, performs O(n log n) in average, worst and best case scenarios even against large sets of data.